+++ tests/test_pipeline.py
@@ -1,169 +1,45 @@
+import string
 from typing import List
 
-import pytest
-from pytest_cases import parametrize_with_cases, parametrize
-
-import time
-import string
-from functools import partial, update_wrapper
-import pandas as pd
 import numpy as np
-
-from sqlalchemy.sql.schema import Column
-from sqlalchemy.sql.sqltypes import String
-from pkg_resources import parse_version
-
-from datapipe.types import data_to_index
-from datapipe.compute import Catalog, Pipeline, Table
+import pandas as pd
+from datapipe.compute import Catalog, Pipeline, Table, build_compute, run_steps
+from datapipe.datatable import DataStore
+from datapipe.step.batch_generate import do_batch_generate
 from datapipe.step.batch_transform import BatchTransform
 from datapipe.step.datatable_transform import DatatableTransformStep
-from datapipe.step.batch_generate import do_batch_generate
-from datapipe.datatable import DataStore
 from datapipe.store.database import TableStoreDB
-from datapipe.compute import build_compute, run_steps
+from datapipe.types import data_to_index
+from label_studio_sdk import LabelStudio
+from pytest_cases import parametrize, parametrize_with_cases
+from sqlalchemy.sql.schema import Column
+from sqlalchemy.sql.sqltypes import JSON, String
 
-from datapipe_label_studio_lite.pipeline import LabelStudioStep
-import label_studio_sdk
-from datapipe_label_studio_lite.sdk_utils import get_project_by_title, is_service_up
-
-
-PROJECT_LABEL_CONFIG_TEST = """<View>
-  <Text name="text" value="$text"/>
-  <Choices name="label" toName="text" choice="single" showInLine="true">
-    <Choice value="Class1"/>
-    <Choice value="Class2"/>
-    <Choice value="Class1_annotation"/>
-    <Choice value="Class2_annotation"/>
-  </Choices>
-</View>"""
-
-
-def wait_until_label_studio_is_up(ls: label_studio_sdk.Client):
-    raise_exception = False
-    counter = 0
-    while not is_service_up(ls, raise_exception=raise_exception):
-        time.sleep(1.0)
-        counter += 1
-        if counter >= 60:
-            raise_exception = True
-
-
-TASKS_COUNT = 10
+from datapipe_label_studio_lite.sdk_utils import get_project_by_title
+from datapipe_label_studio_lite.upload_predictions_pipeline import (
+    LabelStudioUploadPredictions,
+)
+from datapipe_label_studio_lite.upload_tasks_pipeline import LabelStudioUploadTasks
+from tests.ls_test_helpers import (
+    DELETE_UNANNOTATED_TASKS_ONLY_ON_UPDATE,
+    INCLUDE_PARAMS,
+    INCLUDE_PREDICTIONS,
+    PROJECT_LABEL_CONFIG_TEST,
+    TASKS_COUNT,
+    add_predictions,
+    convert_to_ls_input_data,
+    wrapped_partial,
+)
+from tests.util import get_project_id, get_project_tasks, wait_until_label_studio_is_up
 
 
 def gen_data_df():
     yield pd.DataFrame(
         {
             "id": [f"task_{i}" for i in range(TASKS_COUNT)],
-            "text": [
-                np.random.choice([x for x in string.ascii_letters])
-                for i in range(TASKS_COUNT)
-            ],
+            "text": [np.random.choice([x for x in string.ascii_letters]) for i in range(TASKS_COUNT)],
         }
     )
-
-
-def wrapped_partial(func, *args, **kwargs):
-    partial_func = partial(func, *args, **kwargs)
-    update_wrapper(partial_func, func)
-    return partial_func
-
-
-def convert_to_ls_input_data(
-    data_df, include_preannotations: bool, include_prepredictions: bool
-):
-    columns = ["id", "text"]
-
-    for column, bool_ in [
-        ("preannotations", include_preannotations),
-        ("prepredictions", include_prepredictions),
-    ]:
-        if bool_:
-            data_df[column] = [
-                [
-                    {
-                        "result": [
-                            {
-                                "value": {
-                                    "choices": [np.random.choice(["Class1", "Class2"])]
-                                },
-                                "from_name": "label",
-                                "to_name": "text",
-                                "type": "choices",
-                            }
-                        ]
-                    }
-                ]
-                for _ in range(len(data_df))
-            ]
-            columns.append(column)
-
-    return data_df[columns]
-
-
-def add_predictions(data_df):
-    columns = ["id", "predictions", "model_version"]
-    data_df["predictions"] = [
-        {
-            "result": [
-                {
-                    "value": {"choices": [np.random.choice(["Class1", "Class2"])]},
-                    "from_name": "label",
-                    "to_name": "text",
-                    "type": "choices",
-                }
-            ]
-        }
-        for _ in range(len(data_df))
-    ]
-    data_df["model_version"] = "test-model"
-
-    return data_df[columns]
-
-
-INCLUDE_PARAMS = [
-    pytest.param(
-        {"include_preannotations": False, "include_prepredictions": False}, id=""
-    ),
-    # pytest.param(
-    #     {
-    #         'include_preannotations': True,
-    #         'include_prepredictions': False
-    #     },
-    #     id='Preann'
-    # ),
-    # pytest.param(
-    #     {
-    #         'include_preannotations': False,
-    #         'include_prepredictions': True
-    #     },
-    #     id='Prepred'
-    # ),
-    # pytest.param(
-    #     {
-    #         'include_preannotations': True,
-    #         'include_prepredictions': True
-    #     },
-    #     id='PreannPrepred'
-    # ),
-]
-
-INCLUDE_PREDICTIONS = [
-    pytest.param(False, id="NoPredsStep"),
-    # pytest.param(
-    #     True,
-    #     id='WithPredStep'
-    # ),
-]
-
-DELETE_UNANNOTATED_TASKS_ONLY_ON_UPDATE = [
-    pytest.param(False, id="DelAllOnUpdate"),
-    pytest.param(True, id="DelUnAnnOnUpdate"),
-    # pytest.param(
-    #     True,
-    #     id='WithPredStep'
-    # ),
-]
 
 
 class CasesLabelStudio:
@@ -217,57 +93,74 @@ class CasesLabelStudio:
                         create_table=True,
                     )
                 ),
+                "ls_input_data__has__prediction": Table(
+                    store=TableStoreDB(
+                        dbconn=dbconn,
+                        name="ls_input_data__has__prediction",
+                        data_sql_schema=[
+                            Column("id", String(), primary_key=True),
+                            Column("prediction", JSON),
+                        ],
+                        create_table=True,
+                    )
+                ),
             }
         )
-        pipeline = Pipeline(
+        main_steps = [
+            BatchTransform(
+                func=wrapped_partial(
+                    convert_to_ls_input_data,
+                    include_preannotations=include_preannotations,
+                    include_prepredictions=include_prepredictions,
+                    base_columns=["id", "text"],
+                ),
+                inputs=["ls_input_data_raw"],
+                outputs=["ls_input_data"],
+            ),
+            LabelStudioUploadTasks(
+                input__item="ls_input_data",
+                output__label_studio_project_task="ls_task",
+                output__label_studio_project_annotation="ls_output",
+                output__label_studio_sync_table="ls_sync_datetime",
+                ls_url=ls_url,
+                api_key=api_key,
+                project_identifier=project_title,
+                project_label_config_at_create=PROJECT_LABEL_CONFIG_TEST,
+                primary_keys=["id"],
+                columns=["text"],
+                create_table=True,
+                delete_unannotated_tasks_only_on_update=delete_unannotated_tasks_only_on_update,
+            ),
+        ]
+        predictions_steps = (
             [
                 BatchTransform(
-                    func=wrapped_partial(
-                        convert_to_ls_input_data,
-                        include_preannotations=include_preannotations,
-                        include_prepredictions=include_prepredictions,
-                    ),
+                    func=wrapped_partial(add_predictions, base_columns=["id"]),
                     inputs=["ls_input_data_raw"],
-                    outputs=["ls_input_data"],
+                    outputs=["ls_input_data__has__prediction"],
                 ),
-                LabelStudioStep(
-                    input="ls_input_data",
-                    output="ls_output",
-                    sync_table="ls_sync_datetime",
+                LabelStudioUploadPredictions(
+                    input__item__has__prediction="ls_input_data__has__prediction",
+                    input__label_studio_project_task="ls_task",
+                    output__label_studio_project_prediction="ls_prediction",
                     ls_url=ls_url,
                     api_key=api_key,
-                    dbconn=dbconn,
                     project_identifier=project_title,
-                    project_label_config_at_create=PROJECT_LABEL_CONFIG_TEST,
-                    data_sql_schema=[
-                        Column("id", String(), primary_key=True),
-                        Column("text", String()),
-                    ],
+                    primary_keys=["id"],
                     create_table=True,
-                    delete_unannotated_tasks_only_on_update=delete_unannotated_tasks_only_on_update,
                 ),
             ]
+            if include_predictions
+            else []
         )
-        # predictions_step = LabelStudioPredictionsStep(
-        #     input='ls_input_data_raw',
-        #     output='ls_predictions',
-        #     ls_url=ls_url,
-        #     api_key=api_key,
-        #     project_identifier=project_title,
-        #     primary_data_sql_schema=[
-        #         Column('id', String(), primary_key=True),
-        #     ],
-        #     predictions_column='predictions',
-        #     dbconn=dbconn
-        # )
-        # pipeline = Pipeline([main_step] + ([predictions_step] if include_predictions else []))
+        pipeline = Pipeline(main_steps + predictions_steps)
 
         steps = build_compute(ds, catalog, pipeline)
-        label_studio_session = label_studio_sdk.Client(url=ls_url, api_key=api_key)
+        label_studio_session = LabelStudio(base_url=ls_url, api_key=api_key)
         wait_until_label_studio_is_up(label_studio_session)
         project = get_project_by_title(label_studio_session, project_title)
         if project is not None:
-            project.delete_project(project.id)
+            label_studio_session.projects.delete(id=get_project_id(project))
 
         yield (
             ds,
@@ -283,15 +176,10 @@ class CasesLabelStudio:
 
         project = get_project_by_title(label_studio_session, project_title)
         if project is not None:
-            project.delete_project(project.id)
+            label_studio_session.projects.delete(id=get_project_id(project))
 
 
-@parametrize_with_cases(
-    "ds, catalog, steps, project_title, include_preannotations, include_prepredictions, "
-    "include_predictions, label_studio_session, delete_unannotated_tasks_only_on_update",
-    cases=CasesLabelStudio,
-)
-def test_ls_moderation(
+def ls_moderation_base(
     ds: DataStore,
     catalog: Catalog,
     steps: List[DatatableTransformStep],
@@ -299,7 +187,7 @@ def test_ls_moderation(
     include_preannotations: bool,
     include_prepredictions: bool,
     include_predictions: bool,
-    label_studio_session: label_studio_sdk.Client,
+    label_studio_session: LabelStudio,
     delete_unannotated_tasks_only_on_update: bool,
 ):
     # This should be ok (project will be created, but without data)
@@ -313,6 +201,7 @@ def test_ls_moderation(
         output_dts=[ds.get_table("ls_input_data_raw")],
     )
     run_steps(ds, steps)
+    assert len(ds.get_table("ls_task").get_data()) == TASKS_COUNT
     assert len(ds.get_table("ls_output").get_data()) == TASKS_COUNT
 
     # Проверяем проверку на заливку уже размеченных данных
@@ -321,21 +210,20 @@ def test_ls_moderation(
         df_annotation = ds.get_table("ls_output").get_data()
         for idx in df_annotation.index:
             assert len(df_annotation.loc[idx, "annotations"]) == 1
-            assert df_annotation.loc[idx, "annotations"][0]["result"][0]["value"][
-                "choices"
-            ][0] in ["Class1", "Class2"]
+            assert df_annotation.loc[idx, "annotations"][0]["result"][0]["value"]["choices"][0] in ["Class1", "Class2"]
 
     # Person annotation imitation & incremental processing
     project = get_project_by_title(label_studio_session, project_title)
     assert project is not None
+    project_id = get_project_id(project)
 
-    tasks_res = project.get_tasks()
+    tasks_res = get_project_tasks(label_studio_session, project_id)
     assert len(tasks_res) == TASKS_COUNT
 
     run_steps(ds, steps)
 
     # Check that after second run no tasks are leaking
-    tasks_res = project.get_tasks()
+    tasks_res = get_project_tasks(label_studio_session, project_id)
     assert len(tasks_res) == TASKS_COUNT
 
     tasks = np.array(tasks_res)
@@ -355,33 +243,61 @@ def test_ls_moderation(
             for task in tasks[idxs]
         ]
         for task, annotation in zip(tasks[idxs], annotations):
-            label_studio_session.make_request(
-                "POST",
-                f"api/tasks/{task['id']}/annotations/",
-                json=dict(
-                    result=annotation["result"], was_cancelled=False, task_id=task["id"]
-                ),
+            label_studio_session.annotations.create(
+                id=task["id"],
+                result=annotation["result"],
+                was_cancelled=False,
+                task=task["id"],
             )
         run_steps(ds, steps)
-        idxs_df = pd.DataFrame.from_records(
-            {"id": [task["data"]["id"] for task in tasks[idxs]]}
-        )
-        df_annotation = ds.get_table("ls_output").get_data(
-            idx=data_to_index(idxs_df, ["id"])
-        )
+        idxs_df = pd.DataFrame.from_records({"id": [task["data"]["id"] for task in tasks[idxs]]})
+        df_annotation = ds.get_table("ls_output").get_data(idx=data_to_index(idxs_df, ["id"]))
+        if include_predictions:
+            df_prediction = ds.get_table("ls_prediction").get_data(idx=data_to_index(idxs_df, ["id"]))
+            assert len(df_prediction) == len(df_annotation)
+            df_prediction = pd.merge(df_annotation, df_prediction)
         for idx in df_annotation.index:
-            assert len(df_annotation.loc[idx, "annotations"]) == (
-                1 + include_preannotations
+            assert len(df_annotation.loc[idx, "annotations"]) == (1 + include_preannotations)
+            assert df_annotation.loc[idx, "annotations"][0]["result"][0]["value"]["choices"][0] in (
+                ["Class1", "Class2"]
             )
-            assert df_annotation.loc[idx, "annotations"][0]["result"][0]["value"][
-                "choices"
-            ][0] in (["Class1", "Class2"])
             # if include_predictions:
             #     assert len(df_annotation.loc[idx, 'predictions']) == include_prepredictions + include_predictions
             #     if include_prepredictions or include_predictions:
             #         assert df_annotation.loc[idx, 'predictions'][0]['result'][0]['value']['choices'][0] in (
             #             ["Class1", "Class2"]
             #         )
+
+    assert len(ds.get_table("ls_output").get_data()) == 10
+
+
+@parametrize_with_cases(
+    "ds, catalog, steps, project_title, include_preannotations, include_prepredictions, "
+    "include_predictions, label_studio_session, delete_unannotated_tasks_only_on_update",
+    cases=CasesLabelStudio,
+)
+def test_ls_moderation(
+    ds: DataStore,
+    catalog: Catalog,
+    steps: List[DatatableTransformStep],
+    project_title: str,
+    include_preannotations: bool,
+    include_prepredictions: bool,
+    include_predictions: bool,
+    label_studio_session: LabelStudio,
+    delete_unannotated_tasks_only_on_update: bool,
+):
+    ls_moderation_base(
+        ds=ds,
+        catalog=catalog,
+        steps=steps,
+        project_title=project_title,
+        include_preannotations=include_preannotations,
+        include_prepredictions=include_prepredictions,
+        include_predictions=include_predictions,
+        label_studio_session=label_studio_session,
+        delete_unannotated_tasks_only_on_update=delete_unannotated_tasks_only_on_update,
+    )
 
 
 @parametrize_with_cases(
@@ -397,16 +313,13 @@ def test_ls_when_data_is_changed(
     include_preannotations: bool,
     include_prepredictions: bool,
     include_predictions: bool,
-    label_studio_session: label_studio_sdk.Client,
+    label_studio_session: LabelStudio,
     delete_unannotated_tasks_only_on_update: bool,
 ):
     df1 = pd.DataFrame(
         {
             "id": [f"task_{i}" for i in range(TASKS_COUNT)],
-            "text": (
-                ["a", "b", "c", "d", "e", "f", "g", "h", "i", "j"]
-                + ["a"] * (TASKS_COUNT % 10)
-            )
+            "text": (["a", "b", "c", "d", "e", "f", "g", "h", "i", "j"] + ["a"] * (TASKS_COUNT % 10))
             * (TASKS_COUNT // 10),
         }
     )
@@ -414,10 +327,7 @@ def test_ls_when_data_is_changed(
     df2 = pd.DataFrame(
         {
             "id": [f"task_{i}" for i in range(TASKS_COUNT)],
-            "text": (
-                ["A", "B", "C", "d", "E", "f", "G", "h", "I", "j"]
-                + ["a"] * (TASKS_COUNT % 10)
-            )
+            "text": (["A", "B", "C", "d", "E", "f", "G", "h", "I", "j"] + ["a"] * (TASKS_COUNT % 10))
             * (TASKS_COUNT // 10),
         }
     )
@@ -446,11 +356,18 @@ def test_ls_when_data_is_changed(
 
     project = get_project_by_title(label_studio_session, project_title)
     assert project is not None
+    project_id = get_project_id(project)
 
-    tasks = project.get_tasks()
+    tasks = get_project_tasks(label_studio_session, project_id)
     assert len(tasks) == TASKS_COUNT
 
+    assert len(ds.get_table("ls_task").get_data()) == TASKS_COUNT
     df_ls = ds.get_table("ls_output").get_data()
+    assert len(df_ls) == TASKS_COUNT
+    if include_predictions:
+        df_prediction = ds.get_table("ls_prediction").get_data()
+        assert len(df_prediction) == len(df_ls)
+        df_prediction = pd.merge(df_ls, df_prediction)
 
     for idx in df_ls.index:
         # Разметка не должна уйти:
@@ -474,16 +391,13 @@ def test_ls_when_task_is_missing_from_ls(
     include_preannotations: bool,
     include_prepredictions: bool,
     include_predictions: bool,
-    label_studio_session: label_studio_sdk.Client,
+    label_studio_session: LabelStudio,
     delete_unannotated_tasks_only_on_update: bool,
 ):
     df1 = pd.DataFrame(
         {
             "id": [f"task_{i}" for i in range(TASKS_COUNT)],
-            "text": (
-                ["a", "b", "c", "d", "e", "f", "g", "h", "i", "j"]
-                + ["a"] * (TASKS_COUNT % 10)
-            )
+            "text": (["a", "b", "c", "d", "e", "f", "g", "h", "i", "j"] + ["a"] * (TASKS_COUNT % 10))
             * (TASKS_COUNT // 10),
         }
     )
@@ -491,10 +405,7 @@ def test_ls_when_task_is_missing_from_ls(
     df2 = pd.DataFrame(
         {
             "id": [f"task_{i}" for i in range(TASKS_COUNT)],
-            "text": (
-                ["A", "B", "C", "d", "E", "f", "G", "h", "I", "j"]
-                + ["a"] * (TASKS_COUNT % 10)
-            )
+            "text": (["A", "B", "C", "d", "E", "f", "G", "h", "I", "j"] + ["a"] * (TASKS_COUNT % 10))
             * (TASKS_COUNT // 10),
         }
     )
@@ -513,16 +424,6 @@ def test_ls_when_task_is_missing_from_ls(
     )
     run_steps(ds, steps)
 
-    upload_dt = catalog.get_datatable(ds, "ls_input_data_upload")
-    upload_dt.store_chunk(
-        pd.DataFrame(
-            {
-                "id": ["task_0"],
-                "task_id": [-1],
-            }
-        )
-    )
-
     # These steps should delete old tasks and create new tasks
     do_batch_generate(
         func=_gen2,
@@ -533,11 +434,18 @@ def test_ls_when_task_is_missing_from_ls(
 
     project = get_project_by_title(label_studio_session, project_title)
     assert project is not None
+    project_id = get_project_id(project)
 
-    tasks = project.get_tasks()
-    assert len(tasks) == TASKS_COUNT + 1
+    tasks = get_project_tasks(label_studio_session, project_id)
+    assert len(tasks) == TASKS_COUNT
 
+    assert len(ds.get_table("ls_task").get_data()) == TASKS_COUNT
     df_ls = ds.get_table("ls_output").get_data()
+    assert len(df_ls) == TASKS_COUNT
+    if include_predictions:
+        df_prediction = ds.get_table("ls_prediction").get_data()
+        assert len(df_prediction) == len(df_ls)
+        df_prediction = pd.merge(df_ls, df_prediction)
 
     for idx in df_ls.index:
         # Разметка не должна уйти:
@@ -561,23 +469,12 @@ def test_ls_when_some_data_is_deleted(
     include_preannotations: bool,
     include_prepredictions: bool,
     include_predictions: bool,
-    label_studio_session: label_studio_sdk.Client,
+    label_studio_session: LabelStudio,
     delete_unannotated_tasks_only_on_update: bool,
 ):
-    # Skip this test when LS is 1.4.0 and include_preannotations=True, include_prepredictions=False
-    if (
-        include_preannotations
-        and not include_prepredictions
-        and (parse_version(label_studio_session.version) == parse_version("1.4.0"))
-    ):
-        return
     # These steps should upload tasks
     data_df = next(gen_data_df())
-    data_df2 = (
-        data_df.set_index("id")
-        .drop(index=[f"task_{i}" for i in [0, 3, 5, 7, 9]])
-        .reset_index()
-    )
+    data_df2 = data_df.set_index("id").drop(index=[f"task_{i}" for i in [0, 3, 5, 7, 9]]).reset_index()
 
     def _gen():
         yield data_df
@@ -603,15 +500,20 @@ def test_ls_when_some_data_is_deleted(
 
     project = get_project_by_title(label_studio_session, project_title)
     assert project is not None
+    project_id = get_project_id(project)
 
-    tasks = project.get_tasks()
+    tasks = get_project_tasks(label_studio_session, project_id)
     assert len(tasks) == TASKS_COUNT - 5
 
-    df_ls_upload = ds.get_table("ls_input_data_upload").get_data()
-    assert len(df_ls_upload) == TASKS_COUNT - 5
+    df_ls_task = ds.get_table("ls_task").get_data()
+    assert len(df_ls_task) == TASKS_COUNT - 5
 
     df_ls = ds.get_table("ls_output").get_data()
     assert len(df_ls) == TASKS_COUNT - 5
+    if include_predictions:
+        df_prediction = ds.get_table("ls_prediction").get_data()
+        assert len(df_prediction) == TASKS_COUNT - 5
+        df_prediction = pd.merge(df_ls, df_prediction)
 
     for idx in df_ls.index:
         # Разметка не должна уйти:
@@ -635,7 +537,7 @@ def test_ls_specific_updating_scenary(
     include_preannotations: bool,
     include_prepredictions: bool,
     include_predictions: bool,
-    label_studio_session: label_studio_sdk.Client,
+    label_studio_session: LabelStudio,
     delete_unannotated_tasks_only_on_update: bool,
 ):
     df1 = pd.DataFrame(
@@ -668,10 +570,9 @@ def test_ls_specific_updating_scenary(
     # Add 5 annotations
     project = get_project_by_title(label_studio_session, project_title)
     assert project is not None
-    tasks_before = project.get_tasks()
-    tasks_before_sorted = np.array(
-        sorted(tasks_before, key=lambda task: task["data"]["id"])
-    )
+    project_id = get_project_id(project)
+    tasks_before = get_project_tasks(label_studio_session, project_id)
+    tasks_before_sorted = np.array(sorted(tasks_before, key=lambda task: task["data"]["id"]))
     assert len(tasks_before) == 10
     tasks_ids_before = [task["id"] for task in tasks_before]
 
@@ -691,12 +592,11 @@ def test_ls_specific_updating_scenary(
         for task in tasks_before_sorted[[0, 1, 2, 3, 4]]
     ]
     for task, annotation in zip(tasks_before, annotations):
-        label_studio_session.make_request(
-            "POST",
-            f"api/tasks/{task['id']}/annotations/",
-            json=dict(
-                result=annotation["result"], was_cancelled=False, task_id=task["id"]
-            ),
+        label_studio_session.annotations.create(
+            id=task["id"],
+            result=annotation["result"],
+            was_cancelled=False,
+            task=task["id"],
         )
 
     # Получаем текущую полученную разметку
@@ -711,23 +611,24 @@ def test_ls_specific_updating_scenary(
     # Табличка с лейбел студией должна обновиться
     run_steps(ds, steps)
 
-    tasks_after = project.get_tasks()
+    tasks_after = get_project_tasks(label_studio_session, project_id)
     assert len(tasks_after) == 10
 
-    df_ls_upload = ds.get_table("ls_input_data_upload").get_data()
-    assert len(df_ls_upload) == 10
+    df_ls_task = ds.get_table("ls_task").get_data()
+    assert len(df_ls_task) == 10
 
     df_ls = ds.get_table("ls_output").get_data()
     assert len(df_ls) == 10
+    if include_predictions:
+        df_prediction = ds.get_table("ls_prediction").get_data()
+        assert len(df_prediction) == 10
+        df_prediction = pd.merge(df_ls, df_prediction)
 
-    df_ls = pd.merge(df_ls_upload, df_ls)
+    df_ls = pd.merge(df_ls_task, df_ls)
     for idx in df_ls.index:
         if df_ls.loc[idx, "id"] in [f"task_{i}" for i in [0, 1, 2, 6, 7, 8]]:
             # Разметка при обновлении задачи не должна уйти, если delete_unannotated_tasks_only_on_update=False
-            if (
-                df_ls.loc[idx, "id"] in [f"task_{i}" for i in [0, 1, 2]]
-                and delete_unannotated_tasks_only_on_update
-            ):
+            if df_ls.loc[idx, "id"] in [f"task_{i}" for i in [0, 1, 2]] and delete_unannotated_tasks_only_on_update:
                 assert len(df_ls.loc[idx, "annotations"]) > 0
                 assert df_ls.loc[idx, "task_id"] in tasks_ids_before
             else:
@@ -752,7 +653,7 @@ def test_ls_specific_updating_scenary(
     "include_predictions, label_studio_session, delete_unannotated_tasks_only_on_update",
     cases=CasesLabelStudio,
 )
-def test_ls_moderation_with_duplicates_in_ls(
+def test_ls_moderate_then_delete_task(
     ds: DataStore,
     catalog: Catalog,
     steps: List[DatatableTransformStep],
@@ -760,41 +661,32 @@ def test_ls_moderation_with_duplicates_in_ls(
     include_preannotations: bool,
     include_prepredictions: bool,
     include_predictions: bool,
-    label_studio_session: label_studio_sdk.Client,
+    label_studio_session: LabelStudio,
     delete_unannotated_tasks_only_on_update: bool,
 ):
-    # This should be ok (project will be created, but without data)
-    run_steps(ds, steps)
-    run_steps(ds, steps)
-
-    # Загружаем данные для задач в LS во входную таблицу.
-    do_batch_generate(
-        func=gen_data_df,
+    ls_moderation_base(
         ds=ds,
-        output_dts=[ds.get_table("ls_input_data_raw")],
+        catalog=catalog,
+        steps=steps,
+        project_title=project_title,
+        include_preannotations=include_preannotations,
+        include_prepredictions=include_prepredictions,
+        include_predictions=include_predictions,
+        label_studio_session=label_studio_session,
+        delete_unannotated_tasks_only_on_update=delete_unannotated_tasks_only_on_update,
     )
-    
-    # Добавляем дубликаты задач напрямую в проект LS.
-    tasks_duplicates_to_add = [
-        {
-            "data": {
-                "id": "task_1",
-                "text": "task_1_new_text"
-            }
-        },
-        {
-            "data": {
-                "id": "task_2",
-                "text": "task_2_new_text"
-            }
-        }
-    ]
-    project = get_project_by_title(label_studio_session, project_title)
-    project.import_tasks(tasks=tasks_duplicates_to_add)
-
-    # Запускаем трансформацию.
+    ds.get_table("ls_input_data_raw").delete_by_idx(idx=pd.DataFrame({"id": [f"task_{i}" for i in range(5)]}))
     run_steps(ds, steps)
-    
-    # Проверяем количество задач в LS и данных в выходной таблице трубы.
-    assert len(project.get_tasks()) == TASKS_COUNT + len(tasks_duplicates_to_add)
-    assert len(ds.get_table("ls_output").get_data()) == TASKS_COUNT
+    project = get_project_by_title(label_studio_session, project_title)
+    project_id = get_project_id(project)
+    tasks_after = get_project_tasks(label_studio_session, project_id)
+    if delete_unannotated_tasks_only_on_update:
+        assert len(ds.get_table("ls_output").get_data()) == 10
+        assert len(tasks_after) == 10
+        assert len(ds.get_table("ls_task").get_data()) == 10
+        assert len(ds.get_table("ls_output").get_data()) == 10
+    else:
+        assert len(ds.get_table("ls_output").get_data()) == 5
+        assert len(tasks_after) == 5
+        assert len(ds.get_table("ls_task").get_data()) == 5
+        assert len(ds.get_table("ls_output").get_data()) == 5